refactor(workspace): #623 — collapse workspace into src/agent/workspace/
Move legacy workspace state/templates/watcher into src/agent/workspace/
alongside the existing context loader (renamed from workspace.ts to
context.ts). Only consumer of these modules is the agent runtime, so
keeping them adjacent honors layering better than a top-level workspace/.

config-reloader stays in legacy/workspace/ — it's used by the HTTP
adapter and reloads global config, so it belongs with adapter/daemon
infrastructure (to be relocated separately).
