+++ src/agent/workspace/state.test.ts
