+++ src/agent/workspace/state.ts
