+++ src/agent/workspace/template-files/BOOTSTRAP.md
