+++ src/agent/workspace/template-files/SOUL.md
