+++ src/agent/workspace/template-files/SOUL.subagent.md
