+++ src/agent/workspace/template-files/TOOLS.md
