+++ src/agent/workspace/templates.test.ts
