+++ src/agent/workspace/templates.ts
