+++ src/agent/workspace/watcher.test.ts
