+++ src/agent/workspace/watcher.ts
